Update index.Rmd
I think you have done an incredible amount of work and this project looks great so far. I can see you have a few finishing touches to take care of at the end with the results and conclusions, but you have definitely met the criteria for the first draft. Unfortunately though, when I tried to build your website on my own it didn't work.